src/requirements.txt
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.21
celery==3.1.19
cffi==1.5.2
cryptography==1.3.1
Cython==0.24
defusedxml==0.4.1
Django==1.8.6
django-allauth==0.24.1
django-appconf==1.0.1
-e git://github.com/Star2Billing/django-audiofield.git@2754494ed2eae9670a8f1ec06a21eeebba24cc85#egg=django_audiofield
django-awesome-avatar==1.1.3
django-celery==3.1.17
django-crispy-forms==1.5.2
django-embed-video==1.1.0
django-filter==0.13.0
django-formtools==1.0
django-friendship==1.3.1
django-image-cropping==1.0.3
django-model-utils==2.4
django-notifications-hq==1.0
django-tastypie==0.13.3
djangorestframework==3.3.3
easy-thumbnails==2.3
enum34==1.1.3
geopy==1.11.0
idna==2.1
ipaddress==1.0.16
jsonfield==1.0.3
kombu==3.0.29
lxml==3.6.0
Markdown==2.6.6
ndg-httpsclient==0.4.0
oauthlib==1.0.3
Pillow==3.0.0
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.0.0
python-dateutil==2.5.2
python-http-client==1.2.3
python-mimeparse==1.5.1
python-openid==2.2.5
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.5.0
sendgrid==2.2.1
sendgrid-django==2.0.0
six==1.10.0
smtpapi==0.3.1
urllib3==1.12
